"Armor Ignore" and "Toughness" scripts.
"Armor Ignore" is an amount of armor which will be completely ignored at the time of the damage calculation. Same as "Armor Efficiency" but a flat number, not a percentage.
"Toughness" is an amount of armor which cannot be ignored through the "Armor Ignore" and cannot be stripped from the target by "ToArmor" damage.

But I'm not sure if these two scripts work as intended. I can't test them properly.